kt-rag/requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
chromadb>=0.5.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
pytesseract>=0.3.10
Pillow>=10.0.0
chainlit>=1.0.0
python-dotenv>=1.0.0
unstructured>=0.14.0
rich>=13.0.0
openpyxl>=3.1.0
